Refactor : 코드 이동 및 분리
